funcion que sustrae el tiempo agregada en el main

+++ Respaldo/hijomasizquierdo.h
@@ -4,10 +4,10 @@
 struct Node{
     Node* Hmi;
     Node* HD;
-    int Etiqueta;
+    long Etiqueta;
 
 public:
-    Node(Node* HIzq=0,Node* HDer=0,int etique=0)
+    Node(Node* HIzq=0,Node* HDer=0,long etique=0)
         :Hmi(HIzq)
         ,HD(HDer)
         ,Etiqueta(etique)
@@ -21,19 +21,19 @@ class Arbol
 {
 public:
     Node* raiz;
-    int Elementos;
+    long Elementos;
 
 public:
-    Arbol(Node* raiz=0,int elementos=0);
+    Arbol(Node* raiz=0,long elementos=0);
     ~Arbol();
     void Destruir();
     Node* Raiz();
     void DestruirR(Node* nodo);
-    void PonerRaiz(int etiqueta);
-    Node* AgregarHijoI_esimo(Node* padre,int etiqueta,int pos);
+    void PonerRaiz(long etiqueta);
+    Node* AgregarHijoI_esimo(Node* padre, long etiqueta, long pos);
     void Crear();
-    Node *Buscar(int etiqueta);
-    Node* PreOrdenBusqueda(Node* nodo,int etiqueta,Node* buscado);
+    Node *Buscar(long etiqueta);
+    Node* PreOrdenBusqueda(Node* nodo,long etiqueta,Node* buscado);
     Node* Padre(Node* nodo);
     Node* PreOrdenPadre(Node* raiz,Node* buscado);
     Node* HMI(Node* nodo);
@@ -42,11 +42,11 @@ public:
     bool EsHoja(Node*);
     void BorrarHoja(Node* nodo);
     bool isRaiz(Node* nodo);
-    int Etiqueta(Node* nodo);
+    long Etiqueta(Node* nodo);
     void Vaciar();
-    int NumHijos(Node* nodo);
-    int NumElem();
-    void ModificarEtiqueta(Node* nodo,int etiqueta);
+    long NumHijos(Node* nodo);
+    long NumElem();
+    void ModificarEtiqueta(Node* nodo,long etiqueta);
     void MostrarEtiqueta(Node* nodo);
     Node* HMD(Node*);
